# docs/spec/openapi.yaml
openapi: 3.1.0

info:
  title: API de Conversão
  version: 1.0.0
tags:
  - name: Conversões

paths:
  /conversions:
    post:
      summary: Criar uma conversão
      operationId: conversions/create
      tags:
        - Conversões
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - inputFileName
                - inputFileSize
                - outputFileType
              properties:
                inputFileName:
                  type: string
                  description: O nome do arquivo
                  example: file.docx
                inputFileType:
                  type: string
                  description:
                    O tipo original do arquivo; se não fornecido, será inferido a partir da extensão do nome do arquivo
                  example: docx
                inputFileSize:
                  type: integer
                  description: O tamanho do arquivo em bytes
                  example: 1024
                outputFileType:
                  type: string
                  description: O tipo de arquivo desejado para a conversão
                  example: pdf
      responses:
        '202':
          description: Conversão aceita
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversion'
        '400':
          description: Erro de validação
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '5XX':
          description: Erro inesperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'

  /conversions/{conversionId}:
    get:
      summary: Obter os detalhes de uma conversão
      operationId: conversions/getById
      tags:
        - Conversões
      parameters:
        - name: conversionId
          in: path
          description: O identificador da conversão
          required: true
          schema:
            type: string
          example: tz4a98xxat96iws9zmbrgj3a
      responses:
        '200':
          description: Conversão encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversion'
        '404':
          description: Não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'
        '5XX':
          description: Erro inesperado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'

components:
  schemas:
    Conversion:
      type: object
      required:
        - id
        - state
        - inputFileName
        - inputFileType
        - inputFileSize
        - outputFileName
        - outputFileType
        - outputFileSize
        - createdAt
        - completedAt
      properties:
        id:
          type: string
          description: O identificador da conversão
          example: pfh0haxfpzowht3oi213cqos
        state:
          type: string
          description: |
            O estado da conversão:
            - PENDING: conversão pendente;
            - COMPLETED: conversão concluída;
            - ERROR: conversão com erro.
          enum:
            - PENDING
            - COMPLETED
            - ERROR
          example: PENDING
        inputFileName:
          type: string
          description: O nome do arquivo
          example: file.docx
        inputFileType:
          type: string
          description: O tipo original do arquivo
          example: docx
        inputFileSize:
          type: integer
          description: O tamanho do arquivo em bytes
          example: 1024
        outputFileName:
          type: string
          description: O nome do arquivo convertido
          example: file.pdf
        outputFileType:
          type: string
          description: O tipo de arquivo desejado
          example: pdf
        outputFileSize:
          type: integer
          description: O tamanho do arquivo convertido em bytes
          example: 2048
        createdAt:
          type: string
          format: date-time
          description: A data e hora de criação da conversão
          example: '2021-08-01T12:00:00Z'
        completedAt:
          type: string
          format: date-time
          nullable: true
          description: A data e hora da finalização da conversão
          example: '2021-08-01T12:00:00Z'

    ValidationError:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: A mensagem de erro
          example: Validation error
        issues:
          type: array
          description: Os problemas de validação
          items:
            type: object
            properties:
              message:
                type: string
                description: A mensagem de erro
                example: 'Invalid input: expected string, received number'
              code:
                type: string
                description: O código do erro
                example: invalid_type
              path:
                type: array
                description: O caminho do erro
                items:
                  oneOf:
                    - type: string
                    - type: number
                example: ['names', 1]
            additionalProperties: true

    NotFoundError:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: A mensagem de erro
          example: Not found

    InternalServerError:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: A mensagem de erro
          example: Internal server error